move 'nAry' tests to separate file
